+++ src/loaders/AbstractTwoStepLoader.js
@@ -1,0 +1,8 @@
+import AbstractLoader from './AbstractLoader';
+
+export default class AbstractTwoStepLoader extends AbstractLoader {
+  constructor(dataSource, params) {
+    super(params);
+    this.dataSource = dataSource;
+  }
+}