+++ src/loaders/anndata-loaders/CellsZarrLoader.js
@@ -1,21 +1,21 @@
-import BaseAnnDataLoader from './BaseAnnDataLoader';
 import LoaderResult from '../LoaderResult';
+import AbstractTwoStepLoader from '../AbstractTwoStepLoader';
 
 /**
  * Loader for converting zarr into the cell json schema.
  */
-export default class CellsZarrLoader extends BaseAnnDataLoader {
+export default class CellsZarrLoader extends AbstractTwoStepLoader {
   /**
    * Class method for loading spatial cell centroids.
    * @returns {Promise} A promise for an array of tuples/triples for cell centroids.
    */
   loadXy() {
-    const { xy } = this.options || {};
+    const { xy } = (this.options || {});
     if (this.xy) {
       return this.xy;
     }
     if (!this.xy && xy) {
-      this.xy = this.loadNumeric(xy);
+      this.xy = this.dataSource.loadNumeric(xy);
       return this.xy;
     }
     this.xy = Promise.resolve(null);
@@ -32,7 +32,7 @@ export default class CellsZarrLoader extends BaseAnnDataLoader {
       return this.poly;
     }
     if (!this.poly && poly) {
-      this.poly = this.loadNumeric(poly);
+      this.poly = this.dataSource.loadNumeric(poly);
       return this.poly;
     }
     this.poly = Promise.resolve(null);
@@ -52,7 +52,7 @@ export default class CellsZarrLoader extends BaseAnnDataLoader {
       this.mappings = Promise.all(
         Object.keys(mappings).map(async (coordinationName) => {
           const { key } = mappings[coordinationName];
-          return { coordinationName, arr: await this.loadNumeric(key) };
+          return { coordinationName, arr: await this.dataSource.loadNumeric(key) };
         }),
       );
       return this.mappings;
@@ -67,9 +67,9 @@ export default class CellsZarrLoader extends BaseAnnDataLoader {
    * where subarray is a clustering/factor.
    */
   loadFactors() {
-    const { factors } = this.options || {};
+    const { factors } = (this.options || {});
     if (factors) {
-      return this.loadCellSetIds(factors);
+      return this.dataSource.loadObsVariables(factors);
     }
     return Promise.resolve(null);
   }
@@ -80,7 +80,7 @@ export default class CellsZarrLoader extends BaseAnnDataLoader {
         this.loadMappings(),
         this.loadXy(),
         this.loadPoly(),
-        this.loadCellNames(),
+        this.dataSource.loadObsIndex(),
         this.loadFactors(),
       ]).then(([mappings, xy, poly, cellNames, factors]) => {
         const cells = {};