+++ src/loaders/data-sources/ZarrDataSource.js
@@ -1,16 +1,11 @@
 import { HTTPStore, KeyError } from 'zarr';
-import AbstractLoader from './AbstractLoader';
 
 /**
  * A loader ancestor class containing a default constructor
  * and a stub for the required load() method.
  */
-export default class AbstractZarrLoader extends AbstractLoader {
-  constructor(params) {
-    super(params);
-
-    // eslint-disable-next-line no-unused-vars
-    const { url, requestInit } = this;
+export default class ZarrDataSource {
+  constructor({ url, requestInit }) {
     // TODO: We should probably add a way of allowing HEAD requests as well:
     // https://github.com/gzuidhof/zarr.js/blob/375ce0c299469a970da6bb5653513564e25806bb/docs/getting-started/remote-data.md#stores
     const supportedMethods = ['GET'];